+++ ClassLibraries/Macross.Json.Extensions/Code/Properties/AssemblyInfo.cs
@@ -7,4 +7,4 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("46de8812-8b05-40b9-8012-7702f9f31506")]
 
-[assembly: AssemblyVersion("2.0.0.21012")]
+[assembly: AssemblyVersion("2.1.0.21073")]